Organize maker tx overview components
