+++ packages/app/src/features/dialogs/savings/common/components/maker-transaction-overview/index.ts
@@ -1,0 +1,1 @@
+export { MakerTransactionOverview } from './MakerTransactionOverview'
